backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==3.7.1
APScheduler==3.11.0
attrs==25.3.0
black==25.1.0
boto3==1.39.12
botocore==1.39.12
cachetools==6.2.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.5
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
emergent-plugins @ file:///wheelhouse/emergent_plugins-1.0.0-py3-none-any.whl#sha256=c0f704b1f3e22b0bfa65048252154d4524b82184a05034acfd677391e1d02c6e
et_xmlfile==2.0.0
fastapi==0.104.1
flake8==7.3.0
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.26.0
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
h11==0.16.0
httplib2==0.31.0
idna==3.10
iniconfig==2.1.0
isort==6.0.1
jmespath==1.0.1
jq==1.10.0
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
motor==3.3.2
multidict==6.6.4
mypy==1.17.0
mypy_extensions==1.1.0
numpy==2.3.1
oauthlib==3.3.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.8
playwright==1.55.0
pluggy==1.6.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.1
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.5.0
pydantic_core==2.14.1
pyee==13.0.0
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.5.0
pyparsing==3.2.5
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.0.0
rsa==4.9.1
s3transfer==0.13.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.27.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.24.0
watchfiles==1.1.0
xlsxwriter==3.2.5
yarl==1.20.1
